terminate instance before deleting subnet
